:recycle: 이종수: OrderDetails.tsx 파일 내 정의된 인터페이스 새 파일로 분리
